requirements.txt
libs/pgwindow-0.1.1-py3-none-any.whl
pygame-shaders
numpy
pygame_ce